+++ RELEASENOTES.md
@@ -17,6 +17,7 @@
 *   Audio:
     *   Add support for all linear PCM sample formats in
         `ChannelMappingAudioProcessor` and `TrimmingAudioProcessor`.
+    *   Add support for audio gaps in `CompositionPlayer`.
 *   Video:
     *   Improve smooth video frame release at startup when audio samples don't
         start at exactly the requested position.